refactor: simplify reset handling
The name and value on the reset button where there to disambiguate when we posted multiple forms to the same page. No longer needed now that we submit the request to a separate endpoint.

Also no need to respond with any content body, a 205 status header is enough for the frontend to know that it can clear the messages.